+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [v16.0.0] - 2025-09-05
 ### Added
+- Added support for Polish 🇵🇱! This includes the reading comprehension dataset PoQuAD, the sentiment classification dataset PolEmo 2.0, and the linguistic acceptability dataset ScaLA-pl.
 - Added support for Latvian 🇱🇻! This includes the sentiment classification dataset
   Latvian Twitter Sentiment, the linguistic acceptability dataset ScaLA-lv, the named
   entity recognition datasets FullStack-NER-lv and WikiANN-lv, the reading comprehension